supervise: exit when only controlfd gets HUP
We will ignore a HUP on the statusfd. (We'll just silently fail to write
to it in the future.)

This allows pointing the statusfd at, say, a file on the filesystem. And
it allows a single pipe to control the lifetime of the process; when all
write-ends of that pipe are closed, the "reference count" of the process
is 0, and it should be killed.