supervise: don't raise(sig) on fatal signal sig
Instead we'll just kill all our children and go around the main loop
again. This means status messages will be printed. This discards the
pretense of dying from the original signal. Any observer could already
easily see through that by looking at the pid which sent the signal, and
I don't think it was useful for any purpose for us to pretend to die
from the signal.