install dummy SIGCHLD handler to disable SIG_IGN
We are blocking SIGCHLD anyway, so surely we don't need to install a
handler?  Well, if SIGCHLD has been marked as SIG_IGN explicitly that
has the additional effect of disabling zombification of children - they
instead just exit directly.  I'm not sure if that effect is inherited
across exec, but to be safe, let's install an actual handler as a way to
reset this effect.