don't send "no_children" status twice
We do need a separate status message just before termination, to
indicate that we are terminating cleanly. But in a normal case, where
the child processes all exit cleanly without us having to filicide them,
we will also get a "no_children" message from read_childfd. Let's change
termination to send a message which is distinct from that one, so we can
really understand what's going on.