python: call setsid() after forking to stop controlling tty SIGINT
Python code is frequently run from a REPL in a terminal, and normally
a Ctrl-C in the REPL in a terminal will cause SIGINT to be sent to the
entire foreground process group. This means Ctrl-C in the REPL will
not just raise KeyboardInterrupt in Python, it will also kill child
processes. This is inconvenient; and because supervise handles cleanup
of all child processes, it's also not necessary.

Therefore we leave the parent's process group and terminal (by calling
setsid); this means Ctrl-C in the REPL won't kill this subprocess; the
subprocess will only be killed if the fd is closed (which will happen
naturally if Python exits without setting up another supervisor).